move things to correct locations
