+++ types/validator_set_test.go
@@ -18,6 +18,22 @@ import (
 	tmtime "github.com/tendermint/tendermint/types/time"
 )
 
+// RandValidatorSet returns a randomized validator set, useful for testing.
+// NOTE: PrivValidator are in order.
+// UNSTABLE
+func RandValidatorSet(numValidators int, votingPower int64) (*ValidatorSet, []PrivValidator) {
+	valz := make([]*Validator, numValidators)
+	privValidators := make([]PrivValidator, numValidators)
+	for i := 0; i < numValidators; i++ {
+		val, privValidator := RandValidator(false, votingPower)
+		valz[i] = val
+		privValidators[i] = privValidator
+	}
+	vals := NewValidatorSet(valz)
+	sort.Sort(PrivValidatorsByAddress(privValidators))
+	return vals, privValidators
+}
+
 func TestValidatorSetBasic(t *testing.T) {
 	// empty or nil validator lists are allowed,
 	// but attempting to IncrementProposerPriority on them will panic.
